requirements.txt
mypy

faker

requests
types-requests

psycopg2-binary
types-psycopg2

lxml

annotated-types
bcrypt
pandas
